# fate-manager/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "fate manager api documents",
        "title": "Fate-Manager",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:9090",
    "paths": {
...
        "/fate-manager/api/site/checkUrl": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SiteController"
                ],
                "summary": "check register url state",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.RegisterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.BoolResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
        "/fate-manager/api/site/institutions": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SiteController"
                ],
                "summary": "Get Apply List",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.InstitutionsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
        "/fate-manager/api/user/add": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserController"
                ],
                "summary": "Add User",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.AddUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
        "/fate-manager/api/user/delete": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserController"
                ],
                "summary": "Delete User",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.DeleteUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app.BoolResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "boolean"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "app.InstitutionsResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.ApplyResult"
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "app.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "entity.AddUserReq": {
            "type": "object",
            "properties": {
                "creator": {
                    "type": "string"
                },
                "partyId": {
                    "type": "integer"
                },
                "permissionList": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "roleId": {
                    "type": "integer"
                },
                "siteName": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                }
            }
        },
...
        "entity.ApplyResult": {
            "type": "object",
            "properties": {
                "institutions": {
                    "type": "string"
                },
                "status": {
                    "type": "object",
                    "$ref": "#/definitions/entity.IdPair"
                }
            }
        },
...
        "entity.DeleteUserReq": {
            "type": "object",
            "properties": {
                "partyId": {
                    "type": "integer"
                },
                "siteName": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                }
            }
        },
...
        "entity.IdPair": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "desc": {
                    "type": "string"
                }
            }
        },
...
        "entity.RegisterReq": {
            "type": "object",
            "properties": {
                "appKey": {
                    "type": "string"
                },
                "appSecret": {
                    "type": "string"
                },
                "federatedOrganization": {
                    "type": "string"
                },
                "federatedUrl": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "institutions": {
                    "type": "string"
                },
                "networkAccessEntrances": {
                    "type": "string"
                },
                "networkAccessExits": {
                    "type": "string"
                },
                "partyId": {
                    "type": "integer"
                },
                "registrationLink": {
                    "type": "string"
                },
                "role": {
                    "type": "integer"
                },
                "siteName": {
                    "type": "string"
                }
            }
        },
...
    }
}